Added new api for [..]T; updated Static_Array
Trying out a new API for [..]T that feels more like Go/Odin. This
is definitely subject to change, but will at least stay in line
with Basic's for the time being.

Static_Array now has an operator overload for []*.

Union.jai was using tabs, but I'm not fully committed to that yet,
so it's back to spaces (for now).